requirements.txt
beautifulsoup4==4.12.3
Jinja2==3.1.6
markdown==3.8
mcp==1.26.0
Pillow==11.3.0
requests==2.32.5
